+++ DB/db설계.JPG
