+++ UI/다둥다둥 폰트, 색상.txt
@@ -1,0 +1,59 @@
+*시작화면 
+배경색 fd7e6b
+폰트색 ffffff
+폰트 타이포_크레파스 B 
+
+
+*비밀번호 화면
+배경색 ffffff
+폰트색 fd7e6b
+다둥다둥 타이포_크레파스 B 
+숫자폰트 나눔스퀘어 ExtraBold 
+
+
+*회원가입 화면
+배경 ffffff
+흰바탕 글씨 fd7e6b
+버튼 fd7e6b
+버튼 글씨 ffffff
+welcome 타이포_크레파스 B
+나머지 글자들 나눔스퀘어 Regular
+
+
+*메인화면 
+폰트 나눔스퀘어 regular (찾기 - a6a6a6, 
+			 날짜 - 4일째~ :fd7e6b 
+			 2017년 3월 15일 - ffffff) 
+     나눔스퀘어 Bold (오늘 - 104a56 
+		      생일 - ffffff)      
+상단 fd7e6b
+찾기 겉부분 e6e6e6
+안부분 f8f8f8
+찾기 폰트색  a6a6a6
+흰 배경 원과 선, 플러스 버튼 104a56
+작은 원 bebebe
+흰선 ffffff
+ 
+
+*기록(개인)
+모유 f27e7e
+분유 f1b600
+이유식 38cecc
+기저귀 c28d65
+수면 7476b1
+유축 ee8bc1
+유축수유 9ecc07
+목욕 60c7f4
+예방접종 e25423
+기타 b883eb
+상단바 fd7e6b
+
+폰트 나눔스퀘어 Regular
+
+클릭시 색상 ffe4e0
+
+날짜 폰트 08526b
+시간 폰트 fd7e6b
+수정 멘트 484848
+화살표 7b7b7b 
+
