--- sleep_input.xml
@@ -1,186 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin">
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/Stime"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:background="@drawable/border"
-            android:gravity="center"
-            android:paddingBottom="15dp"
-            android:text="시간"
-            android:textColor="#fd7e6d"
-            android:textSize="20dp"
-            android:textStyle="bold" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:paddingTop="5dp">
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginLeft="10dp"
-                android:layout_marginRight="10dp"
-                android:paddingTop="5dp"
-                android:text="수면 시간"
-                android:textColor="#7d7d7d"
-                android:textSize="15dp"
-                android:textStyle="bold" />
-            <TextView
-                android:id="@+id/Shour"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="(시간)"
-                android:textStyle="bold"
-                android:textColor="#fd7d6e"/>
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:orientation="horizontal"
-                android:weightSum="1"
-                android:paddingLeft="5dp"
-                android:paddingTop="10dp"
-                android:paddingBottom="7dp"
-                android:background="@drawable/border2">
-
-                <Button
-                    android:id="@+id/Sm30"
-                    android:layout_width="45dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="-30"
-                    android:textColor="#ffffff"/>
-                <Button
-                    android:id="@+id/Sm5"
-                    android:layout_width="45dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="-5"
-                    android:textColor="#ffffff"/>
-
-                <Button
-                    android:id="@+id/Sm1"
-                    android:layout_width="45dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="-1"
-                    android:textColor="#ffffff"/>
-
-                <EditText
-                    android:id="@+id/SminuteEdit"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:ems="3"
-                    android:text="0"
-                    android:gravity="right"
-                    android:inputType="number" />
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="분"
-                    android:textColor="#fd7d6e"
-                    android:textStyle="bold"
-                    />
-
-                <Button
-                    android:id="@+id/Sp1"
-                    android:layout_width="45dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="+1"
-                    android:textColor="#ffffff"/>
-
-                <Button
-                    android:id="@+id/Sp5"
-                    android:layout_width="50dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="+5"
-                    android:textColor="#ffffff"/>
-                <Button
-                    android:id="@+id/Sp30"
-                    android:layout_width="50dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@color/colorAccent"
-                    android:text="+30"
-                    android:textColor="#ffffff"/>
-            </LinearLayout>
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"
-                android:background="@drawable/border2"
-                android:paddingBottom="10dp">
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginLeft="10dp"
-                    android:layout_marginRight="10dp"
-                    android:paddingTop="5dp"
-                    android:text="수면 시간"
-                    android:textColor="#7d7d7d"
-                    android:textSize="15dp"
-                    android:textStyle="bold" />
-                <RadioGroup
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="horizontal"
-                    android:gravity="center"
-                    android:paddingTop="10dp">
-                    <RadioButton
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="수면"/>
-                    <RadioButton
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:text="낮잠"/>
-                </RadioGroup>
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:gravity="center_horizontal"
-                android:paddingLeft="20dp"
-                android:paddingRight="20dp"
-                android:layout_marginTop="50dp"
-                >
-
-                <Button
-                    android:id="@+id/Sclose"
-                    android:layout_width="150dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="?android:attr/colorBackground"
-                    android:text="취소" />
-
-                <Button
-                    android:id="@+id/Ssave"
-                    android:layout_width="150dp"
-                    android:layout_height="wrap_content"
-                    android:backgroundTint="@android:color/holo_red_light"
-                    android:text="저장"
-                    android:textColor="#ffffff" />
-            </LinearLayout>        </LinearLayout>
-    </LinearLayout>
-
-</RelativeLayout>
